pyproject.toml
[project]
name = "albumentations-mcp"
version = "0.1.0"
description = "MCP-compliant image augmentation server using Albumentations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "mcp>=1.12.3",
    "pillow>=11.3.0",
]
